+++ packages/cli-plugin-dev/test/faas.test.ts
@@ -5,7 +5,7 @@ import * as assert from 'assert';
 const fetch = require('node-fetch');
 const cwd = join(__dirname, 'fixtures/faas');
 describe('test/faas.test.ts', () => {
-  it('dev', async () => {
+  it('dev esbuild', async () => {
     const { close, port } = await run(cwd, { silent: true, fast: 'esbuild' });
     const response = await fetch(`http://127.0.0.1:${port}/hello?name=midway`);
     const body = await response.text();