+++ packages/cli-plugin-swc/js/swc-register.js
@@ -1,0 +1,2 @@
+const { register } = require('@swc-node/register/register');
+register();