+++ packages/serverless-spec-builder/src/fc/builder.ts
@@ -66,7 +66,7 @@ export class FCSpecBuilder extends SpecBuilder {
             funSpec.initializer ||
             handler.split('.').slice(0, -1).join('.') + '.initializer',
           Handler: handler,
-          Runtime: funSpec.runtime || providerData.runtime || 'nodejs12',
+          Runtime: funSpec.runtime || providerData.runtime || 'nodejs14',
           CodeUri: funSpec.codeUri || '.',
           Timeout: funSpec.timeout || providerData.timeout || 3,
           InitializationTimeout:
@@ -283,9 +283,9 @@ export class FCComponentSpecBuilder extends SpecBuilder {
       internetAccess: providerData.internetAccess,
       role: providerData.role,
       logConfig: lowercaseObjectKey(providerData.logConfig),
-      vpc: lowercaseObjectKey(providerData.vpcConfig),
-      nas: lowercaseObjectKey(providerData.nasConfig),
-      tracing: providerData.tracingConfig,
+      vpcConfig: lowercaseObjectKey(providerData.vpcConfig),
+      nasConfig: lowercaseObjectKey(providerData.nasConfig),
+      tracingConfig: providerData.tracingConfig,
     };
     const region = providerData.region;
     const access = (providerData as any).access || 'default';
@@ -315,7 +315,7 @@ export class FCComponentSpecBuilder extends SpecBuilder {
             initializationTimeout:
               funSpec.initTimeout || providerData.initTimeout || 3,
             memorySize: funSpec.memorySize || providerData.memorySize || 128,
-            runtime: funSpec.runtime || providerData.runtime || 'nodejs12',
+            runtime: funSpec.runtime || providerData.runtime || 'nodejs14',
             timeout: funSpec.timeout || providerData.timeout || 3,
             codeUri: funSpec.codeUri || '.',
             instanceConcurrency: funSpec.concurrency || 1,