+++ packages/cli-plugin-aws/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.99-beta.2](https://github.com/midwayjs/cli/compare/v1.2.99-beta.1...v1.2.99-beta.2) (2022-02-16)
+
+**Note:** Version bump only for package @midwayjs/fcli-plugin-aws
+
+
+
+
+
 ## [1.2.97](https://github.com/midwayjs/cli/compare/v1.2.96...v1.2.97) (2022-01-24)
 
 **Note:** Version bump only for package @midwayjs/fcli-plugin-aws